+++ clang/lib/DPCT/CMakeLists.txt
@@ -42,6 +42,9 @@ set(RUNTIME_HEADERS
   ${CMAKE_SOURCE_DIR}/../clang/runtime/dpct-rt/include/dpct/detail/lapack_utils_detail.hpp
   ${CMAKE_SOURCE_DIR}/../clang/runtime/dpct-rt/include/dpct/detail/rng_utils_detail.hpp
   ${CMAKE_SOURCE_DIR}/../clang/runtime/dpct-rt/include/dpct/detail/sparse_utils_detail.hpp
+  ${CMAKE_SOURCE_DIR}/../clang/runtime/dpct-rt/include/dpct/detail/dnnl_utils_detail.hpp
+  ${CMAKE_SOURCE_DIR}/../clang/runtime/dpct-rt/include/dpct/detail/blas_gemm_utils_detail.hpp
+  ${CMAKE_SOURCE_DIR}/../clang/runtime/dpct-rt/include/dpct/detail/ccl_utils_detail.hpp
   )
 
 set(PROCESS_FILES_OUTPUT
@@ -82,6 +85,9 @@ set(PROCESS_FILES_OUTPUT
   ${CMAKE_BINARY_DIR}/tools/clang/include/clang/DPCT/detail/lapack_utils_detail.hpp.inc
   ${CMAKE_BINARY_DIR}/tools/clang/include/clang/DPCT/detail/rng_utils_detail.hpp.inc
   ${CMAKE_BINARY_DIR}/tools/clang/include/clang/DPCT/detail/sparse_utils_detail.hpp.inc
+  ${CMAKE_BINARY_DIR}/tools/clang/include/clang/DPCT/detail/dnnl_utils_detail.hpp.inc
+  ${CMAKE_BINARY_DIR}/tools/clang/include/clang/DPCT/detail/blas_gemm_utils_detail.hpp.inc
+  ${CMAKE_BINARY_DIR}/tools/clang/include/clang/DPCT/detail/ccl_utils_detail.hpp.inc
 )
 
 add_custom_command(
