+++ clang/runtime/dpct-rt/CMakeLists.txt
@@ -31,6 +31,9 @@ include/dpct/detail/blas_utils_detail.hpp
 include/dpct/detail/lapack_utils_detail.hpp
 include/dpct/detail/rng_utils_detail.hpp
 include/dpct/detail/sparse_utils_detail.hpp
+include/dpct/detail/dnnl_utils_detail.hpp
+include/dpct/detail/blas_gemm_utils_detail.hpp
+include/dpct/detail/ccl_utils_detail.hpp
 )
 
 set(dpct_rt_dpstd_files
